+++ packages/react-components/react-carousel/library/src/components/Carousel/useCarousel.ts
@@ -37,6 +37,7 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     announcement,
     motion = 'slide',
     autoplayInterval = 4000,
+    appearance = 'flat',
   } = props;
 
   const { dir } = useFluent();
@@ -131,6 +132,7 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     ),
 
     activeIndex,
+    appearance,
     circular,
     containerRef: mergedContainerRef,
     viewportRef,