+++ packages/react-components/react-carousel/stories/src/Carousel/CarouselAppearance.stories.tsx
@@ -1,0 +1,102 @@
+import * as React from 'react';
+import type { JSXElement } from '@fluentui/react-components';
+import {
+  Button,
+  Carousel,
+  CarouselAnnouncerFunction,
+  CarouselCard,
+  CarouselNav,
+  CarouselNavButton,
+  CarouselNavContainer,
+  CarouselSlider,
+  CarouselViewport,
+  Image,
+  makeStyles,
+  tokens,
+  typographyStyles,
+} from '@fluentui/react-components';
+
+const useClasses = makeStyles({
+  bannerCard: {
+    alignContent: 'center',
+    height: '450px',
+    textAlign: 'left',
+    position: 'relative',
+  },
+  cardContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    gap: '8px',
+    position: 'absolute',
+    left: '10%',
+    top: '25%',
+    background: tokens.colorNeutralBackground1,
+    padding: '18px',
+    maxWidth: '270px',
+    width: '50%',
+  },
+  title: {
+    ...typographyStyles.title1,
+  },
+  subtext: {
+    ...typographyStyles.body1,
+  },
+});
+
+const IMAGES = [
+  'https://fabricweb.azureedge.net/fabric-website/assets/images/swatch-picker/sea-full-img.jpg',
+  'https://fabricweb.azureedge.net/fabric-website/assets/images/swatch-picker/bridge-full-img.jpg',
+  'https://fabricweb.azureedge.net/fabric-website/assets/images/swatch-picker/park-full-img.jpg',
+  'https://fabricweb.azureedge.net/fabric-website/assets/images/swatch-picker/sea-full-img.jpg',
+  'https://fabricweb.azureedge.net/fabric-website/assets/images/swatch-picker/bridge-full-img.jpg',
+  'https://fabricweb.azureedge.net/fabric-website/assets/images/swatch-picker/park-full-img.jpg',
+];
+
+const BannerCard: React.FC<{ children: React.ReactNode; imageSrc: string; index: number }> = props => {
+  const { children, imageSrc, index } = props;
+  const classes = useClasses();
+
+  return (
+    <CarouselCard className={classes.bannerCard} aria-label={`${index + 1} of ${IMAGES.length}`} id={`test-${index}`}>
+      <Image fit="cover" src={imageSrc} role="presentation" />
+
+      <div className={classes.cardContainer}>
+        <div className={classes.title}>{children}</div>
+        <div className={classes.subtext}>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
+          magna aliqua. Ut enim ad minim veniam.
+        </div>
+        <div>
+          <Button size="small" shape="square" appearance="primary">
+            Call to action
+          </Button>
+        </div>
+      </div>
+    </CarouselCard>
+  );
+};
+const getAnnouncement: CarouselAnnouncerFunction = (index: number, totalSlides: number) => {
+  return `Elevated carousel slide ${index + 1} of ${totalSlides}`;
+};
+
+export const Appearance = (): JSXElement => (
+  <Carousel appearance="elevated" groupSize={1} circular announcement={getAnnouncement}>
+    <CarouselViewport>
+      <CarouselSlider>
+        {IMAGES.map((imageSrc, index) => (
+          <BannerCard key={`image-${index}`} imageSrc={imageSrc} index={index}>
+            Card {index + 1}
+          </BannerCard>
+        ))}
+      </CarouselSlider>
+    </CarouselViewport>
+    <CarouselNavContainer
+      layout="inline"
+      autoplayTooltip={{ content: 'Autoplay', relationship: 'label' }}
+      nextTooltip={{ content: 'Go to next', relationship: 'label' }}
+      prevTooltip={{ content: 'Go to prev', relationship: 'label' }}
+    >
+      <CarouselNav>{index => <CarouselNavButton aria-label={`Carousel Nav Button ${index}`} />}</CarouselNav>
+    </CarouselNavContainer>
+  </Carousel>
+);